test/fixtures/wpt/performance-timeline/worker-with-performance-observer.html

<!doctype html>
<html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<script>
async_test(function(t) {
  const worker = new Worker("resources/worker-with-performance-observer.js");
  worker.onmessage = function(event) {
    t.step(() => assert_equals(event.data, 'SUCCESS'));
    t.done();
  }
}, 'Worker: Test Performance Observer inside a worker.');
</script>
</body>
</html>
